refactor: move extension specific tests to separate file
Running `make all` would overwrite the additional tests cases we added for the added append and subtract methods.
